requirements.txt
asgiref==3.5.2
channels==4.0.0
Django==3.2.16
djangorestframework_jwt==1.11.0
drf_yasg==1.21.4
python-decouple==3.8
rest_framework_simplejwt==0.0.1
daphne==4.0.0
django_crontab==0.7.1
psycopg2_binary==2.9.5
django_mirage_field==1.4.0
drf_spectacular==0.26.1